+++ frontend/src/app/coding-test/result/page.tsx
@@ -1,14 +1,16 @@
+"use client";
 import React from "react";
 import Head from "next/head";
-import Header from "../../components/header";
-import Footer from "../../components/Footer";
+import Header from "../../../components/header";
+import Footer from "../../../components/Footer";
 import Link from "next/link";
-import { useRouter } from "next/router";
-import styles from "../../styles/coding-test.module.css";
+import { useRouter } from "next/navigation"; // Changed from next/router
+import styles from "../../../styles/coding-test.module.css";
 
 const CodingTestResultPage: React.FC = () => {
   const router = useRouter();
-  const { id } = router.query;
+  const searchParams = new URLSearchParams(window.location.search);
+  const id = searchParams.get("id");
 
   // 가상의 테스트 결과 데이터
   const testResult = {
