+++ frontend/src/app/signup/page.tsx
@@ -1,43 +1,55 @@
+"use client";
 import React, { useState } from "react";
 import Head from "next/head";
 import Link from "next/link";
-import { useRouter } from "next/router";
-import "../styles/globals.css";
-import styles from "../styles/login.module.css";
+import { useRouter } from "next/navigation"; // Changed from next/router
+
+import styles from "../../styles/signup.module.css";
 import Image from "next/image";
 
-const LoginPage: React.FC = () => {
+const SignupPage: React.FC = () => {
   const router = useRouter();
+  const [name, setName] = useState("");
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
+  const [confirmPassword, setConfirmPassword] = useState("");
   const [error, setError] = useState("");
 
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
     setError("");
 
+    if (password !== confirmPassword) {
+      setError("비밀번호와 비밀번호 확인이 일치하지 않습니다.");
+      return;
+    }
+
     try {
-      // 실제 로그인 로직
-      console.log("로그인 시도:", { email, password });
-      // 로그인 성공 가정
-      router.push("/");
+      // 여기에 실제 회원가입 로직을 구현
+      console.log("회원가입 시도:", { name, email, password });
+
+      // 회원가입 성공 가정
+      router.push("/login");
     } catch (err) {
-      setError(`로그인에 실패했습니다: ${(err as Error).message}`);
+      setError(
+        `회원가입에 실패했습니다: ${
+          err instanceof Error ? err.message : "알 수 없는 오류"
+        }`
+      );
     }
   };
 
   return (
     <div className={styles.container}>
       <Head>
-        <title>로그인 | ALPACO</title>
-        <meta name="description" content="ALPACO 로그인 페이지" />
+        <title>회원가입 | ALPACO</title>
+        <meta name="description" content="ALPACO 회원가입 페이지" />
       </Head>
 
       {/* 상단 로고 영역 */}
       <header className={styles.header}>
         <Link href="/">
           <div className={styles.logo}>
-            {/* 로고 이미지는 /public 폴더 아래에 넣어두고, 경로를 맞춰주세요. */}
             <Image
               src="/alpaco-logo.svg"
               alt="ALPACO"
@@ -56,16 +68,29 @@ const LoginPage: React.FC = () => {
         </Link>
       </header>
 
-      {/* 중앙 메인 로그인 박스 */}
+      {/* 중앙 메인 회원가입 박스 */}
       <main className={styles.main}>
         <div className={styles.formContainer}>
           <h1 className={styles.title}>ALPACO</h1>
-          <h2 className={styles.subtitle}>로그인</h2>
+          <h2 className={styles.subtitle}>회원가입</h2>
           {error && <div className={styles.errorMessage}>{error}</div>}
 
           <form onSubmit={handleSubmit} className={styles.form}>
             <div className={styles.formGroup}>
-              <label htmlFor="email">아이디 (이메일)</label>
+              <label htmlFor="name">이름</label>
+              <input
+                id="name"
+                name="name"
+                type="text"
+                required
+                value={name}
+                onChange={(e) => setName(e.target.value)}
+                placeholder="이름을 입력하세요"
+              />
+            </div>
+
+            <div className={styles.formGroup}>
+              <label htmlFor="email">이메일</label>
               <input
                 id="email"
                 name="email"
@@ -77,13 +102,13 @@ const LoginPage: React.FC = () => {
                 placeholder="이메일을 입력하세요"
               />
             </div>
+
             <div className={styles.formGroup}>
               <label htmlFor="password">비밀번호</label>
               <input
                 id="password"
                 name="password"
                 type="password"
-                autoComplete="current-password"
                 required
                 value={password}
                 onChange={(e) => setPassword(e.target.value)}
@@ -91,22 +116,31 @@ const LoginPage: React.FC = () => {
               />
             </div>
 
-            <button type="submit" className={styles.loginButton}>
-              로그인
-            </button>
-
-            {/* 비밀번호 재설정/찾기 페이지로 이동시키려면 아래를 Link로 변경하거나 onClick 등으로 처리 */}
-            <div className={styles.resetPasswordLink}>
-              <a href="#">비밀번호를 잊으셨나요?</a>
+            <div className={styles.formGroup}>
+              <label htmlFor="confirmPassword">비밀번호 확인</label>
+              <input
+                id="confirmPassword"
+                name="confirmPassword"
+                type="password"
+                required
+                value={confirmPassword}
+                onChange={(e) => setConfirmPassword(e.target.value)}
+                placeholder="비밀번호를 다시 입력하세요"
+              />
             </div>
 
-            <button
-              type="button"
-              className={styles.createAccountButton}
-              onClick={() => router.push("/signup")}
-            >
-              계정 만들기
+            <button type="submit" className={styles.signupButton}>
+              회원가입
             </button>
+
+            <div className={styles.loginPrompt}>
+              <p>
+                이미 계정이 있으신가요?{" "}
+                <Link href="/login" className={styles.loginLink}>
+                  로그인
+                </Link>
+              </p>
+            </div>
           </form>
         </div>
       </main>
@@ -126,4 +160,4 @@ const LoginPage: React.FC = () => {
   );
 };
 
-export default LoginPage;
+export default SignupPage;
