+++ frontend/src/app/storage/page.tsx
@@ -1,3 +1,4 @@
+"use client";
 import React, { useState } from "react";
 import Head from "next/head";
 import Header from "../../components/header";
