+++ frontend/src/components/ProblemExample.tsx
@@ -1,3 +1,4 @@
+"use client";
 import React, { useState } from "react";
 import styles from "../styles/components/ProblemExample.module.css";
