+++ frontend/src/components/header.tsx
@@ -1,4 +1,5 @@
 // components/Header.tsx
+"use client";
 import Link from "next/link";
 import React from "react";
 import Image from "next/image";
